Implement IValidate; singleton PKCounters; +
Require unique New GV values

+++ ASCtoFMConverter/EpisodeDescription.cs
@@ -22,7 +22,7 @@ namespace ASCtoFMConverter
         public override string ToString()
         {
             StringBuilder sb = new StringBuilder();
-            if (GVValue != 0) sb.Append("GV " + ((int)GVValue).ToString("0") + ": ");
+            if (GVValue != 0) sb.Append("GV " + GVValue.ToString("0") + ": ");
             else sb.Append("NoVal: ");
             sb.Append("From " + Start.ToString() + " to " + End.ToString());
             if (Exclude != null)
